1.Add new TestRunner class for automation testing 2.Only wheater's values could accept minus values. For other - Math.abs(textfield.value)

+++ WeatherTemp.java
@@ -7,20 +7,20 @@ public class WeatherTemp {
     protected double faringeith = 0.0;
     protected double kelvin = 0.0; // 04.07.2018
     // in Celcius system
-    protected double showCelcius(int grad_val) {   // test 04.07.2018
+    protected double showCelcius(double grad_val) {   // test 04.07.2018
         //String answer = new String("Grade in Celcius ");
         celcius = (grad_val - conv_val) * dim_val_f_to_c;
         return celcius;
     }
      // in Faringeuth system
-    protected double showFaringeith(int grad_val) {  // test 04.07.2018
+    protected double showFaringeith(double grad_val) {  // test 04.07.2018
         //String answer = new String("Grade in Faringeith ");
         faringeith = grad_val*dim_val_c_to_f + conv_val;
         return faringeith;
     }
 
     // Kelvin to Celcius 04.07.2018
-    protected double showCelciusFromKelvin(int grad_val) {
+    protected double showCelciusFromKelvin(double grad_val) {
         final double final_zero = -273.15;
         if (grad_val == 0) return final_zero;
         final int kelvin_const = 273;  // C = K - 273;
